backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Agent Framework
langgraph==0.0.26
langchain==0.1.6
langchain-openai==0.0.5
langchain-anthropic==0.1.1

# LLM Router
litellm==1.29.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# Vector DB
qdrant-client==1.7.3

# Embeddings
sentence-transformers==2.3.1
torch==2.1.2
transformers==4.37.2

# Code Analysis
tree-sitter==0.21.0
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.0
tree-sitter-typescript==0.21.0

# Docker SDK
docker==7.0.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
pyyaml==6.0.1
jinja2==3.1.3
python-multipart==0.0.9

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring
prometheus-client==0.19.0
python-json-logger==2.0.7

# Development
pytest==8.0.0
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
